Fix bug in calling asciidoctor-pdf on Linux
I never encountered this while developing the script on macOS (using GNU
Bash), however, using Bash 4.4.20(1)-release on Linux Mint 19.2, the
script doesn't work, because asciidoctor-pdf says that it can't read a
FIFO, only a file. Given that, this change is being made to have it run
properly on Bash in Linux.